erfactor: extract navigation types
