+++ src/molecules/navigation-menu/index.tsx
@@ -13,24 +13,7 @@ import {
   NavigationMenuTrigger,
   navigationMenuTriggerStyle,
 } from '@/components/ui/navigation-menu';
-
-export type submenuTypes = {
-  description: string;
-  href: string;
-  title: string;
-};
-
-export type navigationLinkTypes = {
-  href?: string;
-  submenu?: submenuTypes[];
-  text?: string;
-  title?: string;
-};
-
-export type NavigationProps = {
-  className?: string;
-  navigationLinks?: navigationLinkTypes[];
-};
+import { NavigationProps } from './types';
 
 export default function Navigation({
   navigationLinks,
