+++ src/molecules/navigation-menu/types.ts
@@ -1,0 +1,17 @@
+export type submenuTypes = {
+  description: string;
+  href: string;
+  title: string;
+};
+
+export type navigationLinkTypes = {
+  href?: string;
+  submenu?: submenuTypes[];
+  text?: string;
+  title?: string;
+};
+
+export type NavigationProps = {
+  className?: string;
+  navigationLinks?: navigationLinkTypes[];
+};
